[CSHW-3]: Домашнее задание 3

See merge request natasha.buranova/Route256Five!3

+++ src/Ozon.Route256.Five.OrderService/Controllers/DTO/Orders/OrdersInRegionResponse.cs
@@ -3,7 +3,7 @@
 public class OrdersInRegionResponse
 {
     public string? RegionName { get; set; }
-    public long CountOrders { get; set; }
+    public int CountOrders { get; set; }
     public long TotalSumOrders { get; set; }
     public long TotalWeight { get; set; }
     public long CountClients { get; set; }
